requirements.txt
# ==================== Agent 框架 + Web UI ====================
qwen-agent>=0.0.1
gradio>=4.0.0

# ==================== DashScope SDK ====================
dashscope>=1.17.0
openai>=1.0.0

# ==================== 爬虫相关 ====================
playwright>=1.40.0
playwright-stealth>=1.0.6
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
markdownify>=0.12.0

# ==================== 向量数据库 ====================
faiss-cpu>=1.7.4

# ==================== 数据处理 ====================
numpy>=1.24.0
Pillow>=10.0.0

# ==================== 配置 & 工具 ====================
pydantic>=2.5.0
pyyaml>=6.0.1
loguru>=0.7.0
httpx>=0.27.0
tenacity>=8.2.0
fake-useragent>=1.4.0
tqdm>=4.66.0

# ==================== 定时调度（可选） ====================
apscheduler>=3.10.0
